THULLO-55: updated github workflow
In this version, we first check if the HTML file contains the failures class using grep -q '<td class="failures">'. We also check if the value has a percentage sign before setting the output variable using grep -q '%<\/td>'. If both conditions are met, we extract the percentage value using grep -oP '(?<=<td class="failures">)[0-9]+%' and set it as an output variable named percentage. If either condition fails, we output a message indicating that no percentage was found.